// src/app/interfaces/films.swapi.interface.ts
export interface FilmsResponse {
  count: number;
  next?: any;
  previous?: any;
  results: Episode[];
}

export interface Episode {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: string[];
  planets: string[];
  starships: string[];
  vehicles: string[];
  species: string[];
  created: string;
  edited: string;
  url: string;
}